Add endpoint for finding threads with another user
The /threads/find?user=[userid]&private=[0|1] endpoint now
exists to facilitate searching for threads that you are in
with another user. The ?private flag allows you to enable
"strict" searching -- that is, only return threads that are
only occupied by yourself and that other person, without any
third parties. This is for privacy.